+++ src/event_service.ts
@@ -1,34 +1,46 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+import {InternalServerError} from '@essential-projects/errors_ts';
 import {IEventAggregator} from '@essential-projects/event_aggregator_contracts';
 import {IIAMService, IIdentity} from '@essential-projects/iam_contracts';
 import {APIs, DataModels, Messages} from '@process-engine/consumer_api_contracts';
-import {FlowNodeInstance, IFlowNodeInstanceService, IProcessModelUseCases} from '@process-engine/persistence_api.contracts';
+import {
+  FlowNodeInstance,
+  ICorrelationService,
+  IFlowNodeInstanceService,
+  IProcessModelUseCases,
+  Model,
+} from '@process-engine/persistence_api.contracts';
+import {IProcessModelFacade, IProcessModelFacadeFactory} from '@process-engine/process_engine_contracts';
 
-import {EventConverter} from './converters/index';
 import {applyPagination} from './paginator';
+import * as ProcessModelCache from './process_model_cache';
 
 export class EventService implements APIs.IEventConsumerApi {
 
+  private readonly correlationService: ICorrelationService;
   private readonly eventAggregator: IEventAggregator;
-  private readonly eventConverter: EventConverter;
   private readonly flowNodeInstanceService: IFlowNodeInstanceService;
   private readonly iamService: IIAMService;
   private readonly processModelUseCase: IProcessModelUseCases;
+  private readonly processModelFacadeFactory: IProcessModelFacadeFactory;
 
   private readonly canTriggerMessagesClaim = 'can_trigger_messages';
   private readonly canTriggerSignalsClaim = 'can_trigger_signals';
 
   constructor(
+    correlationService: ICorrelationService,
     eventAggregator: IEventAggregator,
     flowNodeInstanceService: IFlowNodeInstanceService,
     iamService: IIAMService,
+    processModelFacadeFactory: IProcessModelFacadeFactory,
     processModelUseCase: IProcessModelUseCases,
-    eventConverter: EventConverter,
   ) {
+    this.correlationService = correlationService;
     this.eventAggregator = eventAggregator;
     this.flowNodeInstanceService = flowNodeInstanceService;
     this.iamService = iamService;
+    this.processModelFacadeFactory = processModelFacadeFactory;
     this.processModelUseCase = processModelUseCase;
-    this.eventConverter = eventConverter;
   }
 
   public async getEventsForProcessModel(
@@ -42,7 +54,7 @@ export class EventService implements APIs.IEventConsumerApi {
 
     const suspendedEvents = suspendedFlowNodeInstances.filter(this.isFlowNodeAnEvent);
 
-    const eventList = await this.eventConverter.convertEvents(identity, suspendedEvents);
+    const eventList = await this.convertFlowNodeInstancesToEvents(identity, suspendedEvents);
 
     // TODO: Remove that useless `EventList` datatype and just return an Array of Events.
     // Goes for the other UseCases as well.
@@ -73,7 +85,7 @@ export class EventService implements APIs.IEventConsumerApi {
       }
     });
 
-    const eventList = await this.eventConverter.convertEvents(identity, accessibleEvents);
+    const eventList = await this.convertFlowNodeInstancesToEvents(identity, accessibleEvents);
 
     eventList.events = applyPagination(eventList.events, offset, limit);
 
@@ -98,7 +110,7 @@ export class EventService implements APIs.IEventConsumerApi {
       return flowNodeIsEvent && flowNodeBelongstoCorrelation;
     });
 
-    const eventList = await this.eventConverter.convertEvents(identity, suspendedEvents);
+    const eventList = await this.convertFlowNodeInstancesToEvents(identity, suspendedEvents);
 
     eventList.events = applyPagination(eventList.events, offset, limit);
 
@@ -125,8 +137,102 @@ export class EventService implements APIs.IEventConsumerApi {
     this.eventAggregator.publish(signalEventName, payload);
   }
 
+  private async convertFlowNodeInstancesToEvents(
+    identity: IIdentity,
+    suspendedFlowNodes: Array<FlowNodeInstance>,
+  ): Promise<DataModels.Events.EventList> {
+
+    const suspendedEvents: Array<DataModels.Events.Event> = [];
+
+    for (const suspendedFlowNode of suspendedFlowNodes) {
+
+      // A triggerable suspended event will always have an eventType attached to it, to indicate what the event is waiting for.
+      // This will be either a signal or a message.
+      const flowNodeIsNotATriggerableEvent = suspendedFlowNode.eventType !== DataModels.Events.EventType.messageEvent
+                                          && suspendedFlowNode.eventType !== DataModels.Events.EventType.signalEvent;
+
+      if (flowNodeIsNotATriggerableEvent) {
+        continue;
+      }
+
+      const processModelFacade = await this.getProcessModelForFlowNodeInstance(identity, suspendedFlowNode);
+
+      const flowNodeModel = processModelFacade.getFlowNodeById(suspendedFlowNode.flowNodeId);
+
+      const event = await this.convertToConsumerApiEvent(flowNodeModel as Model.Events.Event, suspendedFlowNode);
+
+      suspendedEvents.push(event);
+    }
+
+    const eventList: DataModels.Events.EventList = {
+      events: suspendedEvents,
+      totalCount: suspendedEvents.length,
+    };
+
+    return eventList;
+  }
+
   private isFlowNodeAnEvent(flowNodeInstance: FlowNodeInstance): boolean {
     return flowNodeInstance.eventType !== undefined;
+  }
+
+  private async getProcessModelForFlowNodeInstance(
+    identity: IIdentity,
+    flowNodeInstance: FlowNodeInstance,
+  ): Promise<IProcessModelFacade> {
+
+    let processModel: Model.Process;
+
+    // We must store the ProcessModel for each user, to account for lane-restrictions.
+    // Some users may not be able to see some lanes that are visible to others.
+    const cacheKeyToUse = `${flowNodeInstance.processInstanceId}-${identity.userId}`;
+
+    const cacheHasMatchingEntry = ProcessModelCache.hasEntry(cacheKeyToUse);
+    if (cacheHasMatchingEntry) {
+      processModel = ProcessModelCache.get(cacheKeyToUse);
+    } else {
+      const processModelHash = await this.getProcessModelHashForProcessInstance(identity, flowNodeInstance.processInstanceId);
+      processModel = await this.processModelUseCase.getByHash(identity, flowNodeInstance.processModelId, processModelHash);
+      ProcessModelCache.add(cacheKeyToUse, processModel);
+    }
+
+    const processModelFacade = this.processModelFacadeFactory.create(processModel);
+
+    return processModelFacade;
+  }
+
+  private async getProcessModelHashForProcessInstance(identity: IIdentity, processInstanceId: string): Promise<string> {
+    const processInstance = await this.correlationService.getByProcessInstanceId(identity, processInstanceId);
+
+    return processInstance.hash;
+  }
+
+  private convertToConsumerApiEvent(flowNodeModel: Model.Events.Event, suspendedFlowNode: FlowNodeInstance): DataModels.Events.Event {
+
+    const consumerApiEvent: DataModels.Events.Event = {
+      id: suspendedFlowNode.flowNodeId,
+      flowNodeInstanceId: suspendedFlowNode.id,
+      correlationId: suspendedFlowNode.correlationId,
+      processModelId: suspendedFlowNode.processModelId,
+      processInstanceId: suspendedFlowNode.processInstanceId,
+      eventType: <DataModels.Events.EventType> suspendedFlowNode.eventType,
+      eventName: this.getEventDefinitionFromFlowNodeModel(flowNodeModel, suspendedFlowNode.eventType),
+      bpmnType: suspendedFlowNode.flowNodeType,
+    };
+
+    return consumerApiEvent;
+  }
+
+  private getEventDefinitionFromFlowNodeModel(flowNodeModel: Model.Events.Event, eventType: string): string {
+
+    switch (eventType) {
+      case DataModels.Events.EventType.messageEvent:
+        return (flowNodeModel as any).messageEventDefinition.name;
+      case DataModels.Events.EventType.signalEvent:
+        return (flowNodeModel as any).signalEventDefinition.name;
+      default:
+        throw new InternalServerError(`${flowNodeModel.id} is not a triggerable event!`);
+    }
   }
 
 }
